projects/daas-data-lake-sub-dataflow-bigquery-develop/daas-data-lake-sub-dataflow-bigquery-develop/historical/requirements.txt
snowflake-connector-python
google-api-python-client
google-cloud-bigquery
